feat: 걸음수를 통해 칼로리를 계산하는 기능 추가
칼로리의 값이 정수형 타입에 담기에는 값이 너무 작아서 칼로리 값을 정수형타입에서 실수형으로 변경

+++ src/main/java/server/stepmate/mission/dto/MissionDetailDto.java
@@ -12,7 +12,7 @@ public class MissionDetailDto {
 
     private MissionType missionType;
 
-    private Integer currentValue;
+    private double currentValue;
 
     private Integer goal;
 }
